# backend/generated/openapi/src/main/java/com/whipup/generated/api/MeApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.19.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.whipup.generated.api;

import com.whipup.generated.model.AddOwnedIngredientsRequest;
import com.whipup.generated.model.ErrorResponse;
import com.whipup.generated.model.OwnedIngredientListResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-09-23T01:37:49.076375542Z[Etc/UTC]", comments = "Generator version: 7.19.0")
@Validated
@Tag(name = "OwnedIngredient", description = "the OwnedIngredient API")
public interface MeApi {

    default MeApiDelegate getDelegate() {
        return new MeApiDelegate() {};
    }

    String PATH_ADD_OWNED_INGREDIENTS = "/me/ingredients";
    /**
     * POST /me/ingredients
     *
     * @param addOwnedIngredientsRequest  (required)
     * @return Created (status code 201)
     *         or Invalid request (status code 400)
     *         or Resource not found (status code 404)
     *         or Request conflicts with current state (status code 409)
     *         or Authentication failed or missing (status code 401)
     */
    @Operation(
        operationId = "addOwnedIngredients",
        tags = { "OwnedIngredient" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OwnedIngredientListResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Request conflicts with current state", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication failed or missing", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = MeApi.PATH_ADD_OWNED_INGREDIENTS,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<OwnedIngredientListResponse> addOwnedIngredients(
        @Parameter(name = "AddOwnedIngredientsRequest", description = "", required = true) @Valid @RequestBody AddOwnedIngredientsRequest addOwnedIngredientsRequest
    ) {
        return getDelegate().addOwnedIngredients(addOwnedIngredientsRequest);
    }


    String PATH_DELETE_OWNED_INGREDIENT = "/me/ingredients/{userIngredientId}";
    /**
     * DELETE /me/ingredients/{userIngredientId}
     *
     * @param userIngredientId  (required)
     * @return Deleted (status code 204)
     *         or Resource not found (status code 404)
     *         or Authentication failed or missing (status code 401)
     */
    @Operation(
        operationId = "deleteOwnedIngredient",
        tags = { "OwnedIngredient" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Deleted"),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication failed or missing", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = MeApi.PATH_DELETE_OWNED_INGREDIENT,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> deleteOwnedIngredient(
        @NotNull @Parameter(name = "userIngredientId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userIngredientId") Long userIngredientId
    ) {
        return getDelegate().deleteOwnedIngredient(userIngredientId);
    }


    String PATH_GET_OWNED_INGREDIENTS = "/me/ingredients";
    /**
     * GET /me/ingredients
     *
     * @return Owned ingredients (status code 200)
     *         or Authentication failed or missing (status code 401)
     */
    @Operation(
        operationId = "getOwnedIngredients",
        tags = { "OwnedIngredient" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Owned ingredients", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OwnedIngredientListResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication failed or missing", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MeApi.PATH_GET_OWNED_INGREDIENTS,
        produces = { "application/json" }
    )
    default ResponseEntity<OwnedIngredientListResponse> getOwnedIngredients(
        
    ) {
        return getDelegate().getOwnedIngredients();
    }

}