# backend/generated/openapi/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: WhipUp API
  version: 1.0.0
servers:
- url: /api/v1
security:
- bearerAuth: []
paths:
...
  /auth/kakao:
    post:
      operationId: loginWithKakao
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KakaoLoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
          description: Login success
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication failed or missing
      security: []
      tags:
      - Auth
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Auth
...
  /me/ingredients:
    get:
      operationId: getOwnedIngredients
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnedIngredientListResponse"
          description: Owned ingredients
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication failed or missing
      tags:
      - OwnedIngredient
      x-accepts:
      - application/json
      x-tags:
      - tag: OwnedIngredient
    post:
      operationId: addOwnedIngredients
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOwnedIngredientsRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnedIngredientListResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Request conflicts with current state
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication failed or missing
      tags:
      - OwnedIngredient
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: OwnedIngredient
  /me/ingredients/{userIngredientId}:
    delete:
      operationId: deleteOwnedIngredient
      parameters:
      - explode: false
        in: path
        name: userIngredientId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication failed or missing
      tags:
      - OwnedIngredient
      x-accepts:
      - application/json
      x-tags:
      - tag: OwnedIngredient
...
components:
...
  schemas:
...
    KakaoLoginRequest:
      example:
        authorizationCode: authorizationCode
      properties:
        authorizationCode:
          type: string
      required:
      - authorizationCode
      type: object
    LoginResponse:
      example:
        hasOwnedIngredients: true
        accessToken: accessToken
        user:
          userId: 0
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/UserSummary"
        hasOwnedIngredients:
          type: boolean
      required:
      - accessToken
      - hasOwnedIngredients
      - user
      type: object
    UserSummary:
      example:
        userId: 0
      properties:
        userId:
          format: int64
          type: integer
      required:
      - userId
      type: object
...
    OwnedIngredient:
      example:
        ingredientId: 6
        displayName: displayName
        userIngredientId: 0
      properties:
        userIngredientId:
          format: int64
          type: integer
        ingredientId:
          format: int64
          type: integer
        displayName:
          type: string
      required:
      - displayName
      - ingredientId
      - userIngredientId
      type: object
    OwnedIngredientListResponse:
      example:
        items:
        - ingredientId: 6
          displayName: displayName
          userIngredientId: 0
        - ingredientId: 6
          displayName: displayName
          userIngredientId: 0
      properties:
        items:
          items:
            $ref: "#/components/schemas/OwnedIngredient"
          type: array
      required:
      - items
      type: object
    AddOwnedIngredientsRequest:
      example:
        items:
        - ingredientId: 0
        - ingredientId: 0
      properties:
        items:
          items:
            $ref: "#/components/schemas/OwnedIngredientSelection"
          minItems: 1
          type: array
      required:
      - items
      type: object
    OwnedIngredientSelection:
      example:
        ingredientId: 0
      properties:
        ingredientId:
          format: int64
          type: integer
      required:
      - ingredientId
      type: object
...
    ErrorResponse:
      example:
        code: code
        message: message
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http